Use imu-rs to compute heading at 200 Hz
Sampling frequency is implemented as an ESP timer,
and this is observed by using two GPIO pins as flags,
one observing the time taken to acquire the IMU data
and compute the heading, and another that also includes
the time taken to display the heading via the serial console.

A screenshot of a oscilloscope trace and measurement
shows that acquiring+computing fits comfortably within
the 5ms time budget of each timer iteration (for a 200 Hz
sampling frequency).

The other screenshot shows that printing to the console
takes it uncomfortably close to the full 5 ms budget.